+++ client-spark/common/src/main/java/org/apache/uniffle/shuffle/manager/RssShuffleManagerBase.java
@@ -270,7 +270,7 @@ public abstract class RssShuffleManagerBase implements RssShuffleManagerInterfac
   }
 
   /** See static overload of this method. */
-  public abstract long getTaskAttemptIdForBlockId(int mapIndex, int attemptNo);
+  public abstract int getTaskAttemptIdForBlockId(int mapIndex, int attemptNo);
 
   /**
    * Provides a task attempt id to be used in the block id, that is unique for a shuffle stage.
@@ -285,7 +285,7 @@ public abstract class RssShuffleManagerBase implements RssShuffleManagerInterfac
    *
    * @return a task attempt id unique for a shuffle stage
    */
-  protected static long getTaskAttemptIdForBlockId(
+  protected static int getTaskAttemptIdForBlockId(
       int mapIndex, int attemptNo, int maxFailures, boolean speculation, int maxTaskAttemptIdBits) {
     int maxAttemptNo = getMaxAttemptNo(maxFailures, speculation);
     int attemptBits = getAttemptIdBits(maxAttemptNo);
@@ -318,7 +318,7 @@ public abstract class RssShuffleManagerBase implements RssShuffleManagerInterfac
               + "]). Please consider providing more bits for taskAttemptIds.");
     }
 
-    return (long) mapIndex << attemptBits | attemptNo;
+    return mapIndex << attemptBits | attemptNo;
   }
 
   protected static void fetchAndApplyDynamicConf(SparkConf sparkConf) {
